membuat role admin lebih kuat, membuat seeder, README, dan debugging

+++ backend/app/graphql/resolver.ts
@@ -22,11 +22,11 @@ export const root = {
       location: string
       type: string
       date: string
-    } /*context: any*/
+    }, context: any
   ) {
-    // if (context.user?.role !== 'admin') {
-    //     throw new Error('Unauthorized');
-    // }
+    if (context.user?.role !== 'admin') {
+        throw new Error('Unauthorized');
+    }
     const bencana = new Bencana({ title, description, location, type, date })
     return await bencana.save()
   },
@@ -45,21 +45,21 @@ export const root = {
       location: string
       type: string
       date: string
-    } /*context: any*/
+    }, context: any
   ) {
-    // if (context.user?.role !== 'admin') {
-    //     throw new Error('Unauthorized');
-    // }
+    if (context.user?.role !== 'admin') {
+        throw new Error('Unauthorized');
+    }
     return await Bencana.findByIdAndUpdate(
       id,
       { title, description, location, type, date },
       { new: true }
     )
   },
-  async deleteBencana({ id }: { id: string } /*context: any*/) {
-    // if (context.user?.role !== 'admin') {
-    //     throw new Error('Unauthorized');
-    // }
+  async deleteBencana({ id }: { id: string }, context: any) {
+    if (context.user?.role !== 'admin') {
+        throw new Error('Unauthorized');
+    }
     return await Bencana.findByIdAndDelete(id)
   },
 }
